엔티티 추가 생성(JSONB 타입 있는 엔티티는 타입 지정 필요) 및 서비스 레이서 생성

+++ src/main/java/com/hsy/wanted_preonboarding_challenge_backend_31_pre_assignment/api/products/entity/ProductsEntity.java
@@ -15,9 +15,9 @@ import java.time.LocalDateTime;
 public class ProductsEntity {
 
     @Id
-    private int id;
-    private int sellerId;
-    private int brandId;
+    private Long id;
+    private Long sellerId;
+    private Long brandId;
     private String name;
     private String slug;
     private String shortDescription;
